refactor: fixes, shit and more shit ig
