+++ src/api/hooks.ts
@@ -1,6 +1,6 @@
-import { CustomFeatures, CustomGuildInfo } from '@/config/types';
+import { CustomFeatures, CustomGuildInfo } from '../config/types';
 import { QueryClient, useMutation, useQuery } from '@tanstack/react-query';
-import { UserInfo, getGuilds, fetchUserInfo } from '@/api/discord';
+import { getGuild, getGuilds, fetchUserInfo } from '@/api/discord';
 import {
 	disableFeature,
 	enableFeature,
@@ -11,6 +11,7 @@ import {
 	updateFeature,
 } from '@/api/bot';
 import { GuildInfo } from '@/config/types';
+import type { UserInfo } from '@/utils/types';
 import { useAccessToken, useSession } from '@/utils/auth/hooks';
 
 export const client = new QueryClient({
@@ -30,9 +31,21 @@ export const Keys = {
 	login: ['login'],
 	guild_info: (guild: string) => ['guild_info', guild],
 	features: (guild: string, feature: string) => ['feature', guild, feature],
-	guildRoles: (guild: string) => ['guild_roles', guild],
-	guildChannels: (guild: string) => ['guild_channel', guild],
+	guildRoles: (guild: string) => ['gulid_roles', guild],
+	guildChannels: (guild: string) => ['gulid_channel', guild],
 };
+
+export const Mutations = {
+	updateFeature: (guild: string, id: string) => ['feature', guild, id],
+};
+
+export function useGuild(id: string) {
+	const accessToken = useAccessToken();
+
+	return useQuery(['guild', id], () => getGuild(accessToken as string, id), {
+		enabled: accessToken != null,
+	});
+}
 
 export function useGuilds() {
 	const accessToken = useAccessToken();
@@ -92,15 +105,15 @@ export function useEnableFeatureMutation() {
 					if (enabled) {
 						return {
 							...prev,
-							enabledFeatures: prev?.enabledFeatures?.includes(feature)
+							enabledFeatures: prev.enabledFeatures.includes(feature)
 								? prev.enabledFeatures
-								: [...prev.enabledFeatures as string[], feature],
+								: [...prev.enabledFeatures, feature],
 						};
 					}
 					else {
 						return {
 							...prev,
-							enabledFeatures: prev?.enabledFeatures?.filter((f) => f !== feature),
+							enabledFeatures: prev.enabledFeatures.filter((f) => f !== feature),
 						};
 					}
 				});
