+++ src/pages/api/auth/callback.ts
@@ -9,14 +9,30 @@ import { z } from 'zod';
 import { getAbsoluteUrl } from '@/utils/utils';
 
 async function exchangeToken(code: string): Promise<AccessToken> {
+	const data = {
+		client_id: process.env.CLIENT_ID as string,
+		client_secret: process.env.CLIENT_SECRET as string,
+		grant_type: 'authorization_code',
+		code: code,
+		redirect_uri: `${getAbsoluteUrl()}/api/auth/callback`,
+	};
+
+	const headers = {
+		'Content-Type': 'application/x-www-form-urlencoded',
+	};
+
 	const response = await fetch(`${API_ENDPOINT}/oauth2/token`, {
-		headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+		headers,
 		method: 'POST',
-		body: `client_id=${process.env.CLIENT_ID}&client_secret=${process.env.CLIENT_SECRET}&grant_type=authorization_code&code=${code}&redirect_uri=${getAbsoluteUrl()}/api/auth/callback`,
+		body: new URLSearchParams(data),
 	});
 
-	if (response.ok) return (await response.json()) as AccessToken;
-	else throw new Error('Failed to exchange token');
+	if (response.ok) {
+		return (await response.json()) as AccessToken;
+	}
+	else {
+		throw new Error('Failed to exchange token');
+	}
 }
 
 const querySchema = z.object({
@@ -35,11 +51,13 @@ const querySchema = z.object({
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
 	const query = querySchema.safeParse(req.query);
 
-	if (!query.success) return res.status(400).json('Invalid query param');
+	if (!query.success) {
+		return res.status(400).json('Invalid query param');
+	}
 
 	const { code, state } = query.data;
 	const token = await exchangeToken(code);
 
 	setServerSession(req, res, token);
-	return res.redirect(state ? `/${state}/user/home` : '/user/home');
+	res.redirect(state ? `/${state}/user/home` : '/dash/user/home');
 }
