+++ src/pages/api/auth/index.ts
@@ -4,6 +4,10 @@ import { getServerSession } from '@/utils/auth/server';
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
 	const session = getServerSession(req);
 
-	if (!session.success) return res.status(301).redirect('/api/auth/login');
+	if (!session.success) {
+		res.status(401).json('Unauthorized');
+		return;
+	}
+
 	res.status(200).json(session.data);
 }
