+++ src/pages/api/auth/login.ts
@@ -6,5 +6,5 @@ export default function handler(req: NextApiRequest, res: NextApiResponse) {
     locale?: string;
   };
 
-	res.redirect(302, `https://discord.com/oauth2/authorize?client_id=${process.env.CLIENT_ID}&redirect_uri=${getAbsoluteUrl()}/api/auth/callback&response_type=code&scope=identify+guilds&state=${locale ?? ''}`);
+	res.redirect(302, `https://discord.com/oauth2/authorize?client_id=${process.env.CLIENT_ID as string}&redirect_uri=${getAbsoluteUrl()}/api/auth/callback&response_type=code&scope=identify+guilds&scope=identify+guilds&state=${locale ?? ''}`);
 }
