+++ src/pages/api/auth/signout.ts
@@ -3,5 +3,5 @@ import { removeSession } from '@/utils/auth/server';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
 	await removeSession(req, res);
-	return res.status(200).json('logged out');
+	res.status(200).json('logged out');
 }
