+++ src/utils/common.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { CustomFeatures, FeatureConfig } from '@/config/types';
+import { CustomFeatures } from '@/config/types';
 import { features } from '@/config/features';
+import { FeatureConfig } from '@/config/types';
 
 export type IdFeature<K extends keyof CustomFeatures = keyof CustomFeatures> = FeatureConfig<K> & {
   id: K;
@@ -22,3 +23,22 @@ export function toRGB(num: number) {
 		r = (num & 0xff0000) >>> 16;
 	return 'rgb(' + [r, g, b].join(',') + ')';
 }
+
+export const convertHexToRGBA = (hexCode: string, opacity = 1) => {
+	let hex = hexCode.replace('#', '');
+
+	if (hex.length === 3) {
+		hex = `${hex[0]}${hex[0]}${hex[1]}${hex[1]}${hex[2]}${hex[2]}`;
+	}
+
+	const r = parseInt(hex.substring(0, 2), 16);
+	const g = parseInt(hex.substring(2, 4), 16);
+	const b = parseInt(hex.substring(4, 6), 16);
+
+	/* Backward compatibility for whole number based opacity values. */
+	if (opacity > 1 && opacity <= 100) {
+		opacity = opacity / 100;
+	}
+
+	return `rgba(${r},${g},${b},${opacity})`;
+};
